requirements.txt
Flask==3.0.0
balldontlie==0.1.2
python-dotenv==1.0.0